requirements.txt
Flask==2.0.0
gunicorn
python-binance